refactoring (files); created a file for every class
