+++ Rdc-Calculator/Eurozone-Calculator/EuropeanHumanRightsCourt.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+
+namespace Eurozone_Calculator
+{
+    interface IEuropeanHumanRightsCourt
+    {
+        void IsStateRespectingHumanRights();
+    }
+}
