+++ Rdc-calculator/Eurozone-Calculator/Eurozone-Calculator.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>Eurozone_Calculator</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Folder Include="interfaces\" />
+  </ItemGroup>
+
+</Project>
